Adds PFRules module
This commit adds a new module, `PFRules`, that allows Venator to gather
the information in the `/etc/pf.conf` file.

Also added a few improvements on to the repository to include a
`.gitingore` file and a template for the pull requests of the project.
So the pull requests are more descriptive.